Release 17 May 2019

+++ scripts/test-scripts/base_testscript.sh
@@ -1,16 +1,16 @@
 #!/bin/bash
 
-if [ $# != 3 ]
+if [ $# != 4 ]
 then
-	echo "Missing Field, Parameter-list and Config-label."
+	echo "Missing Field, Parameter-list and ConfigFile and DataFile."
 	exit 1
 fi
 
 FIELD=$1
-CONFFILE_ORIG="fastimg_wproj_wenssconfig.json"
-RESFILE_BENCH="res_bench_"$3"_"$FIELD".txt"
-RESFILE1="res_accu_s1_"$3"_"$FIELD".txt"
-RESFILE2="res_accu_s2_"$3"_"$FIELD".txt"
+CONFFILE_ORIG=$3
+MSDATA=$4
+RESFILE_BENCH="res_bench_"$3"_"$4"_"$FIELD".txt"
+RESFILE1="res_accu_"$3"_"$4"_"$FIELD".txt"
 
 echo_run()
 {
@@ -19,9 +19,13 @@ echo_run()
 }
 
 
-echo > $RESFILE_BENCH
-echo > $RESFILE1
-echo > $RESFILE2
+echo -n > $RESFILE_BENCH
+
+if [[ $MSDATA == fivesrcdata* ]] ;
+then
+	echo -n > $RESFILE1
+fi
+
 CONFFILE=tmp_$CONFFILE_ORIG
 cp $CONFFILE_ORIG $CONFFILE
 
@@ -40,20 +44,26 @@ do
 	fi
 	rm logfile.txt
 	i=$(echo "${i//\"}")
-	echo_run "./reduce $CONFFILE wenssdata_awproj.npz -l > /dev/null"
+	echo_run "./reduce $CONFFILE $MSDATA > /dev/null"
 	echo_run "cat logfile.txt | grep ' \[benchmark\] ' | grep ' = ' | awk 'BEGIN {printf \"$i\"}{printf \" %s\", \$NF} END {printf \"\\n\"}' > tmp_$RESFILE_BENCH"
-	echo_run "cat logfile.txt | grep ' \[sources\] ' | grep 'extremum_x_idx=8192, extremum_y_idy=8192' | grep '(?<=extremum_val=).*(?=, extremum_x_idx)' -oP | awk 'BEGIN {printf \"$i\"}{printf \" %s\", \$NF} END {printf \"\\n\"}' > tmp_$RESFILE1"
-	echo_run "cat logfile.txt | grep ' \[sources\] ' | grep 'extremum_x_idx=1259[0-9], extremum_y_idy=340[0-9]' | grep '(?<=extremum_val=).*(?=, extremum_x_idx)' -oP | awk 'BEGIN {printf \"$i\"}{printf \" %s\", \$NF} END {printf \"\\n\"}' > tmp_$RESFILE2"
-	
-    cat $RESFILE_BENCH tmp_$RESFILE_BENCH > out_$RESFILE_BENCH
+	if [[ $MSDATA == fivesrcdata* ]] ;
+	then
+		echo_run "cat logfile.txt | grep ' \[sources\] ' | grep 'Island 0:' | grep '(?<=extremum_val=).*(?=, extremum_x_idx)' -oP | awk 'BEGIN {printf \"$i\"}{printf \" %s\", \$NF} END {printf \" \"}' > tmp_$RESFILE1"
+		echo_run "cat logfile.txt | grep ' \[sources\] ' | grep 'Island 1:' | grep '(?<=extremum_val=).*(?=, extremum_x_idx)' -oP | awk 'BEGIN {}{printf \" %s\", \$NF} END {printf \" \"}' >> tmp_$RESFILE1"
+		echo_run "cat logfile.txt | grep ' \[sources\] ' | grep 'Island 2:' | grep '(?<=extremum_val=).*(?=, extremum_x_idx)' -oP | awk 'BEGIN {}{printf \" %s\", \$NF} END {printf \" \"}' >> tmp_$RESFILE1"
+		echo_run "cat logfile.txt | grep ' \[sources\] ' | grep 'Island 3:' | grep '(?<=extremum_val=).*(?=, extremum_x_idx)' -oP | awk 'BEGIN {}{printf \" %s\", \$NF} END {printf \" \"}' >> tmp_$RESFILE1"
+		echo_run "cat logfile.txt | grep ' \[sources\] ' | grep 'Island 4:' | grep '(?<=extremum_val=).*(?=, extremum_x_idx)' -oP | awk 'BEGIN {}{printf \" %s\", \$NF} END {printf \"\\n\"}' >> tmp_$RESFILE1"
+	fi	
+
+        cat $RESFILE_BENCH tmp_$RESFILE_BENCH > out_$RESFILE_BENCH
 	mv out_$RESFILE_BENCH $RESFILE_BENCH
 	rm tmp_$RESFILE_BENCH
-	cat $RESFILE1 tmp_$RESFILE1 > out_$RESFILE1
-	mv out_$RESFILE1 $RESFILE1
-	rm tmp_$RESFILE1
-	cat $RESFILE2 tmp_$RESFILE2 > out_$RESFILE2
-	mv out_$RESFILE2 $RESFILE2
-	rm tmp_$RESFILE2
+	if [[ $MSDATA == fivesrcdata* ]] ;
+	then
+		cat $RESFILE1 tmp_$RESFILE1 > out_$RESFILE1
+		mv out_$RESFILE1 $RESFILE1
+		rm tmp_$RESFILE1
+	fi
 done
 
 rm $CONFFILE
